refactor: `SummarySection` as its own component
