+++ packages/frontend/src/components/Modals/ReportsModal/ReportsModal.tsx
@@ -13,9 +13,8 @@ import { Report } from 'src/utils/types'
 import { getLineColor } from 'src/utils/uiUtils'
 
 import { FeedbackForm } from '../../Form/FeedbackForm/FeedbackForm'
-import { Line } from '../../Miscellaneous/Line/Line'
-import { ClusteredReportItem } from './ClusteredReportItem'
 import { ReportItem } from './ReportItem'
+import { SummarySection } from './SummarySection'
 
 interface ReportsModalProps {
     className?: string
@@ -216,10 +215,6 @@ const ReportsModal: React.FC<ReportsModalProps> = ({ className, onCloseModal }) 
         [sortedLinesWithReports]
     )
 
-    const riskLevels = [3, 2, 1, 0]
-
-    const filterRiskLevelLines = (level: number, riskData: LineRiskData): boolean => riskData.class === level
-
     if (showFeedback) {
         return <FeedbackForm openAnimationClass={className} />
     }
@@ -239,57 +234,12 @@ const ReportsModal: React.FC<ReportsModalProps> = ({ className, onCloseModal }) 
                 ))}
             </section>
             {currentTab === 'summary' ? (
-                <section className="summary">
-                    <section className="lines">
-                        <div className="align-child-on-line">
-                            <h2>{t('ReportsModal.reportsHeading')}</h2>
-                            <FeedbackButton handleButtonClick={() => setShowFeedback(true)} />
-                        </div>
-                        <p className="time-range">{t('ReportsModal.past24Hours')}</p>
-                        {Array.from(sortedLinesWithReports.entries())
-                            .sort(([, inspectorsA], [, inspectorsB]) => {
-                                const timestampA = new Date(inspectorsA[0].timestamp).getTime()
-                                const timestampB = new Date(inspectorsB[0].timestamp).getTime()
-
-                                return timestampB - timestampA // most recent first
-                            })
-                            .slice(0, 5)
-                            .map(([line, inspectors]) => (
-                                <ClusteredReportItem key={line} inspectors={inspectors} />
-                            ))}
-                    </section>
-                    <section className="risk">
-                        <h2>{t('ReportsModal.risk')}</h2>
-                        <div className="risk-grid">
-                            {riskLevels.map((level) => {
-                                const linesWithRiskLevel = Array.from(riskLines.entries()).filter(([, riskData]) =>
-                                    filterRiskLevelLines(level, riskData)
-                                )
-
-                                if (linesWithRiskLevel.length === 0) return null
-
-                                return (
-                                    <div key={level} className="risk-grid-item">
-                                        {linesWithRiskLevel.map(([line, riskData]) => (
-                                            // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
-                                            <div
-                                                key={line}
-                                                className={`risk-line risk-level-${riskData.class}`}
-                                                onClick={() => onCloseModal()}
-                                            >
-                                                <img
-                                                    src={`/icons/risk-${riskData.class}.svg`}
-                                                    alt="Icon to show risk level"
-                                                />
-                                                <Line line={line} />
-                                            </div>
-                                        ))}
-                                    </div>
-                                )
-                            })}
-                        </div>
-                    </section>
-                </section>
+                <SummarySection
+                    sortedLinesWithReports={sortedLinesWithReports}
+                    riskLines={riskLines}
+                    onCloseModal={onCloseModal}
+                    setShowFeedback={setShowFeedback}
+                />
             ) : null}
             {currentTab === 'lines' ? (
                 <section className="list-modal">
